% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxsa.R
\docType{methods}
\name{diagnostics-method}
\alias{diagnostics-method}
\alias{diagnostics}
\alias{diagnostics,FLXSA-method}
\title{XSA diagnostics}
\usage{
\S4method{diagnostics}{FLXSA}(object, sections = rep(T, 8), ...)
}
\description{
Provides the diagnostics table used in ICES WG to analyse the results of the
XSA run.
}
\section{Generic Function}{
 \describe{ \item{call}{diagnostics(object, ...}
\item{object}{An object with the results of a VPA method.} }
}

\keyword{methods}
