\name{diagnostics-method}
\docType{methods}
\alias{diagnostics}
\alias{diagnostics-method}
\alias{diagnostics,FLXSA-method}
\title{XSA diagnostics}
\description{
	Provides the diagnostics table used in ICES WG to analyse the results of the XSA run.
}
\section{Generic Function}{
\describe{
	\item{call}{diagnostics(object, ...}
	\item{object}{An object with the results of a VPA method.}
}
}
\section{Methods}{
\describe{
\item{object="FLXSA"}{This method will provide a text file similar to the diagnostics output of the VPA lowestoft.}
}}
\keyword{methods}
