% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxsa.R
\docType{class}
\name{FLXSA-class}
\alias{FLXSA-class}
\alias{show,FLXSA-method}
\title{Class FLXSA}
\description{
A class for the results of an XSA analysis.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("FLXSA", ...)} and are output by calls to \code{\link{FLXSA}}.
}

\references{
Darby, C. D., and Flatman, S. 1994. Virtual Population Analysis:
version 3.1 (Windows/Dos) user guide. Info. Tech. Ser., MAFF Direct. Fish.
Res., Lowestoft, (1): 85pp.

Shepherd, J.G. 1992. Extended survivors analysis: an improved method for the
analysis of catch-at-age data and catch-per-unit-effort data. Working paper
No. 11 ICES Multi-species Assessment Working Group, June 1992, Copenhagen,
Denmark. 22pp. (mimeo).

Shepherd, J.G. 1994. Prediction of yearclass strength by calibration
regression analysis of multiple recruit index series. ICES J. Mar. Sci. In
Prep.
}
\seealso{
\code{\link{FLXSA}}
}
\author{
Laurence Kell
}
\keyword{classes}
