% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\docType{class}
\name{FLXSA.control}
\alias{FLXSA.control}
\alias{FLXSA.control-class}
\title{A class for FLXSA control options}
\usage{
FLXSA.control(x = NULL, tol = 1e-09, maxit = 30, min.nse = 0.3,
  fse = 0.5, rage = 0, qage = 10, shk.n = TRUE, shk.f = TRUE,
  shk.yrs = 5, shk.ages = 5, window = 100, tsrange = 20,
  tspower = 3, vpa = FALSE)
}
\description{
Runs of the FLXSA method require a number of parameter to be set. Most of them
change the behaviour of the solution-searching algorithm
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. 
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id
}
\section{Slots}{

\describe{
\item{\code{x}}{An object of class FLXSA. If provided, the 'FLXSA.control' is
initialized with the corresponding values of an XSA analysis stored in the
object. This is useful for getting the same initial sloteters for
successive analyses. Specifying one or more of the other arguments
supersedes default values, or values obtained from this FLXSA object}

\item{\code{tol}}{The covergence tolerance, i.e. difference between two successive
iterations must be lower, to declare convergence of the model.}

\item{\code{maxit}}{The maximum number of iterations allowed}

\item{\code{min.nse}}{The minimum value of SE permitted in estimate of N hat}

\item{\code{fse}}{User set SE of F when shrinking to mean F}

\item{\code{rage}}{The oldest age for which the two parameter model is used for
catchability at age. Note that this value should be one less than the value
used in the executable version of XSA}

\item{\code{qage}}{The age after which catchability is no longer estimated. q at
older ages set to the value at this age}

\item{\code{shk.n}}{If \code{TRUE}, shrinkage to mean N}

\item{\code{shk.f}}{If \code{TRUE}, shrinkage to mean F}

\item{\code{shk.yrs}}{The number of years to be used for shrinkage to F for
terminal year}

\item{\code{shk.ages}}{The number of ages to be used for shrinkage to F for
terminal age}

\item{\code{window}}{The time window to consider in the model}

\item{\code{tsrange}}{The number of years to be used in the time series weighting}

\item{\code{tspower}}{The power to be used in the time series taper weighting}

\item{\code{vpa}}{If \code{FALSE} use cohort analysis, otherwise, use VPA}
}}

\section{Validity}{


  \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLCatch'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity.
}

\examples{

	# To create a new FLXSA.control object with default parameters:
	my.xsa.control <- FLXSA.control()
	my.xsa.control
	# Same, but changing values of some parameters
	my.xsa.control <- FLXSA.control(maxit=50, shk.f=FALSE)
	my.xsa.control

}
\seealso{
\link{FLComp}
}
\author{
Laurence Kell & Philippe Grosjean
}
\keyword{classes}
