% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\docType{methods}
\name{rand}
\alias{rand}
\title{rand}
\arguments{
\item{n}{number of iters to generate}

\item{mean}{an \code{FLXSA} object}
}
\value{
sd an \code{FLStock} object
}
\description{
Uses the internal standard errors to conduct a Monte Carlo simulation of the
terminal N-at-age in the last year
}
\examples{

\dontrun{
data(ple4)
data(ple4.indices)
xsa =FLXSA(ple4,ple4.indices)
ple4=rand(100,ple4,xsa)}

}
