\name{summary-method}
\docType{methods}
\alias{summary,FLXSA-method}
\title{Summary for XSA}
\description{
	Provides the summary table used in ICES WG for XSA.
}
\section{Generic Function}{see See \code{\link{summary}}}
\section{Methods}{
\describe{
\item{object="FLXSA", stock="FLStock", R.age="character", Fbar.rng="character"}{This will output a data.frame with the summary information. Note that the R.age and Fbar.rng arguments are necessary to compute extract R and compute Fbar}
}}
\keyword{methods}
