% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsa.sa.R
\name{xsa.sa}
\alias{xsa.sa}
\title{An FLXSA estimator module for the mse package}
\usage{
xsa.sa(stk, idx, ...)
}
\arguments{
\item{stk}{An FLStock.}

\item{idx}{An FLIndices.}
}
\value{
A list containing the updated FLStock, and the tracking FLQuant.
}
\description{
The `mse` package can use different stock assessment methods as modules in
the estimation step of a management procedure. This functions provides such
a module for FLXSA.
As a flag for convergence, the returned tracking FLQuant contains the number
of maximum number of iterations (maxit).
}
